requirements.txt
snowflake-connector-python==3.12.2
pandas==1.4.3
argparse==1.4.0
APScheduler==3.9.1
numpy==1.26.4
python-dotenv==1.0.1
duckdb==1.1.1
pyyaml==6.0.2